tui: reset keyboard reporting on exit
